data type implementation fixed

+++ template.c
@@ -5,27 +5,27 @@
 #include <limits.h>
 #include <math.h>
 
-int32_t vmax(int32_t vector2[], int32_t n) {
-	int32_t vmax = INT_MIN;
-	for (int32_t i = 0; i < n; i++) {
+int64_t vmax(int64_t vector2[], int64_t n) {
+	int64_t vmax = INT_MIN;
+	for (int64_t i = 0; i < n; i++) {
 		if (vmax < vector2[i]) {
 			vmax = vector2[i];
 		}
 	}
 	return vmax;
 }
-int32_t vmin(int32_t vector2[], int32_t n) {
-	int32_t vmin = INT_MAX;
-	for (int32_t i = 0; i < n; i++) {
+int64_t vmin(int64_t vector2[], int64_t n) {
+	int64_t vmin = INT_MAX;
+	for (int64_t i = 0; i < n; i++) {
 		if (vmin > vector2[i]) {
 			vmin = vector2[i];
 		}
 	}
 	return vmin;
 }
-int32_t vmin_non_zero(int32_t vector2[], int32_t n) {
-	int32_t vmin_non_zero = INT_MAX;
-	for (int32_t i = 0; i < n; i++) {
+int64_t vmin_non_zero(int64_t vector2[], int64_t n) {
+	int64_t vmin_non_zero = INT_MAX;
+	for (int64_t i = 0; i < n; i++) {
 		if (vmin_non_zero > vector2[i] && vector2[i] > 0) {
 			vmin_non_zero = vector2[i];
 		}
@@ -33,9 +33,9 @@ int32_t vmin_non_zero(int32_t vector2[], int32_t n) {
 	return vmin_non_zero;
 }
 
-int8_t *parse_vector2(int32_t n) {
-	int8_t *vector2 = malloc(sizeof(int8_t));
-	int8_t length = (int8_t) floor(log10(n));
+int16_t *parse_vector2(int64_t n) {
+	int16_t *vector2 = malloc(sizeof(int16_t));
+	int16_t length = (int16_t) floor(log10(n));
 	for (int i = 0; i < length; i++) {
 		vector2[length - i] = n % 2;
 		n = n / 10;
